% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{methods}
\name{exampledata_VDYP7}
\alias{exampledata_VDYP7}
\title{The function is to generate example data for VDYP7inR package}
\usage{
exampledata_VDYP7()
}
\value{
a list contains a polygon file and layer file
}
\description{
The function is to generate example data for VDYP7inR package
}
\author{
Yong Luo
}
