% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBYTable.R
\docType{methods}
\name{extractBYTable}
\alias{extractBYTable}
\title{The function is to extract yield tables when biomass is turned on}
\usage{
extractBYTable(rawYieldTable, volIncluded, biomIncluded)
}
\arguments{
\item{rawYieldTable}{data.frame or data.table, The yield table with biomass is turned on.}

\item{volIncluded}{logical, Specifies whether the volume yield table is turned on.}

\item{biomIncluded}{logical, Specifies whether the biomass yield table is turned on.}
}
\value{
biomass yield table, and volume yield table is volume is turned on
}
\description{
The function is to extract yield tables when biomass is turned on
}
\author{
Yingbing Chen, Yong Luo
}
