% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat2TwoFiles.R
\docType{methods}
\name{flat2TwoFiles}
\alias{flat2TwoFiles}
\title{The function is to organize flat input (one input file) and return two input files}
\usage{
flat2TwoFiles(flatFile)
}
\arguments{
\item{flatFile}{data.frame or data.table, The flat file.}
}
\value{
a list two files layFile and polygonFile, which are two input files
for VDYP7 projection
}
\description{
This function is to organize flat input (one input file)
and return two input files
}
\note{
the flatFile only contains primary layer and
veteran layer
}
\author{
Yingbing Chen, Yong Luo
}
