% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilTableGenerator.R
\docType{methods}
\name{utilTableGenerator}
\alias{utilTableGenerator}
\title{The function is to prepare the util input table for runing VDYP7}
\usage{
utilTableGenerator(
  all,
  exception = NULL,
  biomass = FALSE,
  askForConfirm = TRUE
)
}
\arguments{
\item{all}{numeric, Assign all the 16 species group with this DBH util level.}

\item{exception}{character, Modify the util level by species using this argument. This argument
is formed as \code{SP0,util}, for example, \code{AC,4}.}

\item{biomass}{logical, Generate biomass unil table, default is \code{FALSE}. When it is TRUE,
the function ignores inputs from \code{all} and \code{exception} arguments,
as VDYP7 only uses this util when simulate biomass.}

\item{askForConfirm}{logical, Print the output to confirm the util table, default is TRUE.}
}
\value{
util table for 16 species group
}
\description{
The function is to prepare the util input table for runing VDYP7. This function is
to allow DBH setting for each of the 16 possible species group (SP0).
The DBH util level can be supplied from 4, 7.5, 12.5, 17.5 and 22.5.
}
\examples{
\dontrun{
  utiltable1 <- utilTableGenerator(all = 12.5, askForConfirm = FALSE)
  print(utiltable1)
      SP0 util
   1:  AC 12.5
   2:  AT 12.5
   3:   B 12.5
   4:   C 12.5
   5:   D 12.5
   6:   E 12.5
   7:   F 12.5
   8:   H 12.5
   9:   L 12.5
  10:  MB 12.5
  11:  PA 12.5
  12:  PL 12.5
  13:  PW 12.5
  14:  PY 12.5
  15:   S 12.5
  16:   Y 12.5

  utiltable2 <- utilTableGenerator(all = 12.5,
                          exception = c("AC,4", "H, 17.5", "Y, 22.5"),
                          askForConfirm = FALSE)
  print(utiltable2)
      SP0 util
   1:  AC  4.0
   2:  AT 12.5
   3:   B 12.5
   4:   C 12.5
   5:   D 12.5
   6:   E 12.5
   7:   F 12.5
   8:   H 17.5
   9:   L 12.5
  10:  MB 12.5
  11:  PA 12.5
  12:  PL 12.5
  13:  PW 12.5
  14:  PY 12.5
  15:   S 12.5
  16:   Y 22.5

  utiltable3 <- utilTableGenerator(all = 12.5,
                          exception = c("AC,4", "H, 17.5", "Y, 22.5"),
                          biomass = TRUE,
                          askForConfirm = FALSE)
   print()
           SP0 util
        1:  AC 12.5
        2:  AT 12.5
        3:   B 17.5
        4:   C 17.5
        5:   D 12.5
        6:   E 12.5
        7:   F 17.5
        8:   H 17.5
        9:   L 12.5
       10:  MB 12.5
       11:  PA 12.5
       12:  PL 12.5
       13:  PW 12.5
       14:  PY 12.5
       15:   S 17.5
       16:   Y 17.5

}

}
\author{
Yong Luo
when the projBiomass = T, the unitilization level will be overwritten. And the volume and biomass
summaries will be derived using the new unitilization level.
AC   --> 12.5
AT   --> 12.5
B    --> 17.5
C    --> 17.5
D    --> 12.5

\if{html}{\out{<div class="sourceCode">}}\preformatted{       E    --> 12.5
       F    --> 17.5
       H    --> 17.5
       L    --> 12.5
       MB   --> 12.5

       PA   --> 12.5
       PL   --> 12.5
       PW   --> 12.5

       PY   --> 12.5
       S    --> 17.5
       Y    --> 17.5
}\if{html}{\out{</div>}}
}
