% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validInput.R
\docType{methods}
\name{validInput}
\alias{validInput}
\title{The function is to examine whether the data is valid for VDYP7 run}
\usage{
validInput(polyFile, layerFile)
}
\arguments{
\item{polyFile}{data.frame or data.table, The poly file for VDYP7 run.}

\item{layerFile}{data.frame or data.table, The layer file for VDYP7 run.}
}
\value{
a text file to present the results
}
\description{
The VDYP7 takes two input files poly file and layer file. This function is to
examine whether those two input files are valid for VDYP7 run, in terms of
1) data structure and 2) data matchness
}
\author{
Yong Luo
}
