% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDYP7Save.R
\docType{methods}
\name{VDYP7Save}
\alias{VDYP7Save}
\title{The function is to examine whether the data is valid for VDYP7 run}
\usage{
VDYP7Save(
  simuOutput,
  savePath,
  saveName,
  saveFmt = "txt",
  overWrite = FALSE,
  metadata = FALSE
)
}
\arguments{
\item{simuOutput}{list, The output from \code{\link{VDYP7Run}}.}

\item{savePath}{character, Specifies the path to save outputs.}

\item{saveName}{character, Specifies the name of saved outputs.
This is prefix for saved files. This arguement has to be specified.}

\item{saveFmt}{character, Specifies the format you want to save the yield table. Currently, it
supports \code{rds}, \code{txt}, \code{csv} and \code{xlsx}. Default is \code{txt}.
Please note that the rest of the outputs are saved as \code{txt} file.}

\item{overWrite}{logical, Specifies whether the user wish to overwrite the existing file
in the save path. Default is \code{FALSE}.}

\item{metadata}{logical, Specifies whether the user wish to save metadata.
Default is \code{FALSE}.}
}
\value{
no value
}
\description{
The VDYP7 takes two input files poly file and layer file. This function is to
examine whether those two input files are valid for VDYP7 run, in terms of
1) data structure and 2) data matchness
}
\author{
Yong Luo
}
