% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDYP8Run.R
\docType{methods}
\name{VDYP8Run}
\alias{VDYP8Run}
\title{The function is to run VDYP8 in R environment}
\usage{
VDYP8Run(
  polyFile,
  layerFile,
  utilTable,
  BaseRESTAPIURL = "https://nr-vdyp-test-backend.apps.silver.devops.gov.bc.ca/api/v8/",
  timeSeries,
  startTime,
  endTime,
  timeIncrement = 10,
  backward = TRUE,
  forward = TRUE,
  allowBatphsub = FALSE,
  yTableIncProjMode = TRUE,
  yTableIncAge = TRUE,
  yTableIncYear = TRUE,
  yTableIncPolyID = TRUE,
  forceYear = NULL,
  forceRefYear = TRUE,
  forceCrtYear = FALSE,
  secondarySpcHt = FALSE,
  projByLayer = TRUE,
  projBySpecies = TRUE,
  projVolume = TRUE,
  projBiomass = FALSE,
  logFile = FALSE,
  saveIntermediate = FALSE
)
}
\arguments{
\item{polyFile}{data.frame or data.table, The poly file for VDYP8 run.}

\item{layerFile}{data.frame or data.table, The layer file for VDYP8 run.}

\item{utilTable}{data.frame or data.table,
The table contains utilization level for 16 species groups. It can be generated
using \code{\link{utilTableGenerator}}.}

\item{BaseRESTAPIURL}{character, Specifies where your you are accessing VDYP8 REST API.}

\item{timeSeries}{character, Specifies time series for the simulation from
either \code{age}, \code{year} or \code{NULL}.
If \code{NULL}, forceYear must be specified.}

\item{startTime}{integer, Specifies the time to start simulation.}

\item{endTime}{integer, Specifies the time to terminate simulation.}

\item{timeIncrement}{integer, Specifies the time interval to summary simulation results.
Default is \code{10} years.}

\item{backward}{logical, Whether want to do backward simulation. Default is \code{TRUE}.}

\item{forward}{logical, Whether want to do forward simulation. Default is \code{TRUE}.}

\item{allowBatphsub}{logical, Whether to allow Batphsub. (what is this), Default is \code{FALSE}.}

\item{yTableIncProjMode}{logical, Whether want to use proj mode to run simulation. Default is \code{TRUE}.}

\item{yTableIncAge}{logical, Whether want to include age column in yield table. Default is \code{TRUE}.}

\item{yTableIncYear}{logical, Whether want to include Year column in yield table. Default is \code{TRUE}.}

\item{yTableIncPolyID}{logical, Whether yield table include polygon id. Default is \code{TRUE}.}

\item{forceYear}{numeric, Specifies the force year, so that it can be included in yield table.}

\item{forceRefYear}{logical, Whether want to include reference year in yield table. Default is \code{TRUE}.}

\item{forceCrtYear}{logical, Whether want to include current year in yield table. Default is \code{FALSE}.}

\item{secondarySpcHt}{logical, Whether want to project secondary species height in yield table. Default is \code{FALSE}.}

\item{projByLayer}{logical, Whether allow projection summarized by layer???. Default is \code{TRUE}.}

\item{projBySpecies}{logical, Whether allow projection summarized by species. Default is \code{TRUE}.}

\item{projVolume}{logical, whether allow the output include volume. Default is \code{TRUE}.}

\item{projBiomass}{logical, whether allow the output include biomass. Default is \code{FALSE}.}

\item{logFile}{logical, whether output meta data. Default is \code{FALSE}.}

\item{saveIntermediate}{logical, whether output the intermediate files from VDYP projection Default is \code{FALSE}.}
}
\value{
a list that contains
1) simulated stand yield table (equivalent to \code{-o} in VDYP7 console);
2) processing message (equivalent to \code{-e} in VDYP7 console);
3) log message (equivalent to \code{-l} in VDYP7 console);
4) may contain log message from core module (equivalent to \code{-v7log} in VDYP7 console);
5) metadata to run the simulation (cmd file and console specifications)
}
\description{
This function is to prepare two input files (\code{polyFile} and \code{layerFile});
to prepare VDYP configuration and cmd file; to run VDYP8; and load VDYP8 output into
R environment.
}
\note{
\enumerate{
\item when the projBiomass = T, the unitilization level will be overwritten. And the volume and biomass
summaries will be derived using the new unitilization level.
AC   --> 12.5
AT   --> 12.5
B    --> 17.5
C    --> 17.5
D    --> 12.5
E    --> 12.5
F    --> 17.5
H    --> 17.5
L    --> 12.5
MB   --> 12.5
PA   --> 12.5
PL   --> 12.5
PW   --> 12.5
PY   --> 12.5
S    --> 17.5
Y    --> 17.5
}
}
\seealso{
see \code{\link{VDYP7Save}} to save results; \code{\link{VDYP7RunParallel}} to run VDYP7 parallel.
see \code{\link{VDYP7Debug}} to turn on debug mode.
}
\author{
Peter Minter
}
